Added some changes. Not complete
